LayoutTests/fast/css/color-computed-contrast-color-function.html

<!DOCTYPE html>
<html>
<head>
<title>Test CSS Color Module 5 contrast-color() parsing</title>
<meta charset="utf-8">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../imported/w3c/web-platform-tests/css/support/parsing-testcommon.js"></script>
<script src="../../imported/w3c/web-platform-tests/css/support/computed-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 10px;
    color: pink;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>

// Test basic legacy colors
test_computed_value("background-color", "contrast-color(white)", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(white max)", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(black)", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(black max)", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(pink)", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(pink max)", "rgb(0, 0, 0)");

// Test non-legacy colors
test_computed_value("background-color", "contrast-color(color(srgb 1 0 1 / 0.5))", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(color(srgb 1 0 1 / 0.5) max)", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(lab(0.2 0.5 0.2))", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(lab(0.2 0.5 0.2) max)", "rgb(255, 255, 255)");

// Test out-of-gamut colors
test_computed_value("background-color", "contrast-color(color(srgb 10 10 10))", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(color(srgb 10 10 10) max)", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(color(srgb -10 -10 -10))", "rgb(0, 0, 0)");
test_computed_value("background-color", "contrast-color(color(srgb -10 -10 -10) max)", "rgb(0, 0, 0)");

// Test nested contrast-color
test_computed_value("background-color", "contrast-color(contrast-color(pink))", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(contrast-color(pink max))", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(contrast-color(pink) max)", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(contrast-color(pink max) max)", "rgb(255, 255, 255)");

// Test currentColor
test_computed_value("background-color", "contrast-color(currentcolor)", "rgb(0, 0, 0)"); // currentColor is 'pink', so output should be the same as current-color(pink). 
test_computed_value("background-color", "contrast-color(currentcolor max)", "rgb(0, 0, 0)"); // currentColor is 'pink', so output should be the same as current-color(pink). 

// Test color using calc().
test_computed_value("background-color", "contrast-color(color(srgb calc(1 + (sign(20cqw - 10px) * 1)) calc(1 + (sign(20cqw - 10px) * 1)) calc(1 + (sign(20cqw - 10px) * 1))))", "rgb(255, 255, 255)");
test_computed_value("background-color", "contrast-color(color(srgb calc(1 + (sign(20cqw - 10px) * 1)) calc(1 + (sign(20cqw - 10px) * 1)) calc(1 + (sign(20cqw - 10px) * 1))) max)", "rgb(255, 255, 255)");

</script>
</body>
</html>
